[feat] 컴포넌트 나누기1
